Move IndexData to separate module
