+++ src/main/scala/se/lu/nateko/cp/meta/services/sparql/magic/IndexHandler.scala
@@ -32,13 +32,15 @@ import se.lu.nateko.cp.meta.services.sparql.index.FileSizeHierarchicalBitmap.Lon
 import se.lu.nateko.cp.meta.services.sparql.index.Property
 import se.lu.nateko.cp.meta.services.sparql.index.SamplingHeightHierarchicalBitmap
 import se.lu.nateko.cp.meta.services.sparql.index.StringHierarchicalBitmap
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.DataEndGeo
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.DataStartGeo
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.FileNameGeo
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.IndexData
+import se.lu.nateko.cp.meta.services.sparql.magic.index.DataEndGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.DataStartGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.FileNameGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.IndexData
 import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.ObjEntry
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.SubmEndGeo
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.SubmStartGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.SubmEndGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.SubmStartGeo
+import se.lu.nateko.cp.meta.services.sparql.magic.index.StatKey
+import se.lu.nateko.cp.meta.services.sparql.magic.index.StatEntry
 import se.lu.nateko.cp.meta.utils.async.throttle
 import se.lu.nateko.cp.meta.utils.rdf4j.===
 import se.lu.nateko.cp.meta.utils.rdf4j.Rdf4jStatement
