+++ src/test/scala/se/lu/nateko/cp/meta/test/services/sparql/index/SerializationTests.scala
@@ -30,7 +30,7 @@ import se.lu.nateko.cp.meta.services.sparql.index.Property
 import se.lu.nateko.cp.meta.services.sparql.index.Station
 import se.lu.nateko.cp.meta.services.sparql.index.SubmissionEnd
 import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex
-import se.lu.nateko.cp.meta.services.sparql.magic.CpIndex.IndexData
+import se.lu.nateko.cp.meta.services.sparql.magic.index.IndexData
 import se.lu.nateko.cp.meta.services.sparql.magic.IndexHandler
 import se.lu.nateko.cp.meta.utils.rdf4j.Loading
 import se.lu.nateko.cp.meta.utils.rdf4j.accessEagerly
